alef_classroom/requirements.txt
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0,<4.0.0
python-dotenv>=1.0.0,<2.0.0
Pillow>=10.0.0
django-cors-headers>=4.0.0,<5.0.0
whitenoise>=6.0.0,<7.0.0
